Phase B7: extract chat handler into src/commands/chat.rs
