+++ crates/defra-agent-cli/src/commands/mod.rs
@@ -1,4 +1,5 @@
 // Subcommand handler implementations. Populated by Phases B4-B7.
+pub(crate) mod chat;
 pub(crate) mod config;
 pub(crate) mod init;
 pub(crate) mod p2p;
